+++ test/mix_test_watch_test.exs
@@ -1,7 +1,7 @@
 defmodule Mix.Tasks.Test.WatchTest do
   use ExUnit.Case
 
-  @delays 0..29 |> Enum.map(fn n -> n * 2 end)
+  @delays 0..14 |> Enum.map(fn n -> n * 2 end)
 
   for delay <- @delays do
     test "A dot after #{delay} ms" do