+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Changelog
 =========
 
-## v0.2.2
+## v0.2 - 2015 Summer
 
-Tests now run once immediately after running the mix task. (Johan Lind @behe)
+* Tests now run once immediately after running the mix task. (Johan Lind @behe)
+* The `_build` directory is ignored, as well as `deps/`.
+* Erlang `.erl` files are now watched.